# api/openapi.yaml
openapi: 3.0.3
info:
  title: smelt HTTP API
  description: |
    Local HTTP API served by `smelt serve`. Stores decision programs in the
    history database, renders them to shell scripts on demand, and executes
    them as background subprocesses tracked as runs.

    This file is the contract for the HTTP API exactly as `--help` is the
    contract for the CLI: the server routing and models are generated from it.
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:7700
    description: Default local listen address
paths:
...
  /api/programs:
    get:
      operationId: listPrograms
      summary: List all stored programs
      responses:
        "200":
          description: All stored programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createProgram
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Program created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/programs/{id}:
    parameters:
      - $ref: "#/components/parameters/ProgramID"
    get:
      operationId: getProgram
      summary: Get a single program
      responses:
        "200":
          description: The program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateProgram
      summary: Replace a program's fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteProgram
      summary: Delete a program
      responses:
        "204":
          description: Program deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/programs/{id}/run:
    parameters:
      - $ref: "#/components/parameters/ProgramID"
    post:
      operationId: runProgram
      summary: Render and execute the program as a background run
      responses:
        "202":
          description: Run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/runs:
    get:
      operationId: listRuns
      summary: List recent runs
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 50
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [running, done, failed]
      responses:
        "200":
          description: Recent runs, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/runs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Run ID (UUID)
        schema:
          type: string
    get:
      operationId: getRun
      summary: Get a run with its full job list
      responses:
        "200":
          description: The run and its jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: cancelRun
      summary: Cancel a running subprocess (SIGTERM to its process group)
      responses:
        "204":
          description: Run cancelled
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Run exists but has no live process (already finished or not started by this server)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ProgramID:
      name: id
      in: path
      required: true
      description: Program ID (UUID)
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Cond:
      type: object
      description: |
        A single per-file condition. Not every field/op combination is valid:
        string fields (codec, audio, ext) accept only eq/ne; numeric fields
        (height, width, bitrate, duration) accept the full comparison set.
      required: [field, op, value]
      x-go-type: workflow.Cond
      x-go-type-import:
        path: github.com/Raina-Hardik/smelt/internal/workflow
      properties:
        field:
          type: string
          enum: [codec, audio, ext, height, width, bitrate, duration]
        op:
          type: string
          enum: [eq, ne, gt, lt, ge, le]
        value:
          type: string
    Action:
      type: object
      description: What to do with a file when all conditions match.
      required: [cmd]
      x-go-type: workflow.Action
      x-go-type-import:
        path: github.com/Raina-Hardik/smelt/internal/workflow
      properties:
        cmd:
          type: string
          enum: [transcode, check, skip]
        args:
          type: array
          description: Extra CLI flags passed to the action's smelt command.
          items:
            type: string
    Rule:
      type: object
      description: Ordered decision rule — first rule whose conditions all match wins.
      required: [do]
      x-go-type: workflow.Rule
      x-go-type-import:
        path: github.com/Raina-Hardik/smelt/internal/workflow
      properties:
        when:
          type: array
          items:
            $ref: "#/components/schemas/Cond"
        do:
          $ref: "#/components/schemas/Action"
    ProgramInput:
      type: object
      required: [src]
      properties:
        name:
          type: string
        schedule:
          type: string
          description: Optional cron expression.
        src:
          type: string
          description: Root directory the program scans.
        ext:
          type: array
          description: File extensions to match. Defaults to mkv, mp4, avi when empty.
          items:
            type: string
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
    Program:
      type: object
      required: [id, name, schedule, src, ext, rules, created_at, updated_at]
      properties:
        id:
          type: string
          x-go-name: ID
        name:
          type: string
        schedule:
          type: string
        src:
          type: string
        ext:
          type: array
          items:
            type: string
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Run:
      type: object
      required: [run_id, name, status, total, ok, failed, skipped, started_at]
      properties:
        run_id:
          type: string
          x-go-name: RunID
        name:
          type: string
        status:
          type: string
          enum: [running, done, failed, cancelled]
        total:
          type: integer
        ok:
          type: integer
          x-go-name: OK
        failed:
          type: integer
        skipped:
          type: integer
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    Job:
      type: object
      required: [id, source_path, status, pct, updated_at]
      properties:
        id:
          type: integer
          format: int64
          x-go-name: ID
        source_path:
          type: string
        status:
          type: string
        pct:
          type: number
          format: double
        speed:
          type: string
          x-go-type-skip-optional-pointer: true
        updated_at:
          type: string
          format: date-time
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          required: [jobs]
          properties:
            jobs:
              type: array
              items:
                $ref: "#/components/schemas/Job"
    RunAccepted:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
          x-go-name: RunID